review: bound the customer info retry, cover the external-controller recompute
The retry ran unbounded on the caller's thread, which can be main. It's
now a single rebuild: a write landing during it is vanishingly rare and
corrected by that write's own publish.

Adds the first test to exercise the external-purchase-controller branch —
the suite's own container uses the internal controller, so the recompute
had no coverage. Splits PublishedSubscriptionStatus into its own file to
stay under the file-length limit.
